+++ react/OrderPlaced.tsx
@@ -10,6 +10,7 @@ import { branch, renderComponent } from 'recompose'
 
 import { Helmet, withRuntimeContext, ExtensionPoint } from 'vtex.render-runtime'
 import { Button } from 'vtex.styleguide'
+import { usePWA } from 'vtex.store-resources/PWAContext'
 
 import AnalyticsWrapper from './Analytics'
 import Header from './components/Header'
@@ -54,7 +55,8 @@ const OrderPlaced: FunctionComponent<Props & InjectedIntlProps> = ({
   intl,
 }) => {
   const { orderGroup } = orderGroupQuery
-
+  const { settings = {} } = usePWA() || {}
+  const { promptOnCustomEvent } = settings
   return (
     <CurrencyContext.Provider
       value={orderGroup.orders[0].storePreferencesData.currencyCode}>
@@ -68,7 +70,7 @@ const OrderPlaced: FunctionComponent<Props & InjectedIntlProps> = ({
         profile={orderGroup.orders[0].clientProfileData}
         inStore={inStore}
       />
-      <main>
+      <main className="mv6 w-80-ns w-90 center">
         {orderGroup.orders.map((order: Order, index: number) => (
           <OrderInfo
             order={order}
@@ -78,6 +80,9 @@ const OrderPlaced: FunctionComponent<Props & InjectedIntlProps> = ({
             key={order.orderId}
           />
         ))}
+        {promptOnCustomEvent === 'checkout' && (
+          <ExtensionPoint id="promotion-banner" />
+        )}
       </main>
     </CurrencyContext.Provider>
   )